Log Entries are now inserted with the corresponding parameters
The detail view includes a list of the parameters used in the method call - this is still somewhat manual, so I will need to find ways to simplify it.

+++ app/src/main/java/com/scriptient/rxplorer/async/ParameterTableFetchAsyncTask.java
@@ -13,9 +13,9 @@ public class ParameterTableFetchAsyncTask extends AsyncTask<Void, Void, List<Log
 
     private WeakReference<View> viewWeakReference;
 
-    private int logEntryId;
+    private Long logEntryId;
 
-    public ParameterTableFetchAsyncTask( View view, int logEntryId ) {
+    public ParameterTableFetchAsyncTask( View view, Long logEntryId ) {
 
         this.logEntryId = logEntryId;
         viewWeakReference = new WeakReference<>( view );
